Rename layer.rs to layer_content_metadata.rs
