+++ src/data/layer_content_metadata.rs
